plugins/thesis-writing-assistant/tools/requirements.txt
ebooklib>=0.18
beautifulsoup4>=4.12
lxml>=4.9
markdownify>=0.11
readability-lxml>=0.8
